+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -21,6 +21,7 @@ import mozilla.components.concept.engine.webextension.WebExtension
 import org.json.JSONObject
 import org.mozilla.geckoview.GeckoResult
 import org.mozilla.geckoview.GeckoRuntime
+import org.mozilla.geckoview.GeckoRuntimeSettings
 import org.mozilla.geckoview.GeckoSession
 import org.mozilla.geckoview.GeckoWebExecutor
 import org.mozilla.geckoview.WebExtension as GeckoWebExtension
@@ -152,6 +153,16 @@ class GeckoEngine(
         override var preferredColorScheme: PreferredColorScheme
             get() = PreferredColorScheme.from(runtime.settings.preferredColorScheme)
             set(value) { runtime.settings.preferredColorScheme = value.toGeckoValue() }
+
+        override var allowAutoplayMedia: Boolean
+            get() = runtime.settings.autoplayDefault == GeckoRuntimeSettings.AUTOPLAY_DEFAULT_ALLOWED
+            set(value) {
+                runtime.settings.autoplayDefault = if (value) {
+                    GeckoRuntimeSettings.AUTOPLAY_DEFAULT_ALLOWED
+                } else {
+                    GeckoRuntimeSettings.AUTOPLAY_DEFAULT_BLOCKED
+                }
+            }
     }.apply {
         defaultSettings?.let {
             this.javascriptEnabled = it.javascriptEnabled
@@ -163,6 +174,7 @@ class GeckoEngine(
             this.testingModeEnabled = it.testingModeEnabled
             this.userAgentString = it.userAgentString
             this.preferredColorScheme = it.preferredColorScheme
+            this.allowAutoplayMedia = it.allowAutoplayMedia
         }
     }
 }