+++ mobile/android/android-components/docs/changelog.md
@@ -22,6 +22,12 @@ permalink: /changelog/
      information.
    * `Glean.sendPings` has been added for sending custom pings.
 
+* **concept-engine**
+  * Add boolean `allowAutoplayMedia` setting.
+
+* **browser-engine-gecko-nightly**
+  * Implement `allowAutoplayMedia` in terms of `autoplayDefault`.
+
 # 0.50.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.49.0...v0.50.0)